requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
boto3==1.40.9
botocore==1.40.9
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
dill==0.4.0
distro==1.9.0
docker==7.1.0
eval_type_backport==0.2.2
fastapi==0.116.1
filelock==3.19.1
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
mdurl==0.1.2
mock==4.0.3
multidict==6.6.4
multiprocess==0.70.18
numpy==1.26.4
omegaconf==2.3.0
openai==1.99.9
packaging==24.2
pandas==2.3.1
pathos==0.3.4
pillow==11.3.0
platformdirs==4.3.8
pox==0.3.6
ppft==1.7.7
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
s3transfer==0.13.1
sagemaker==2.250.0
sagemaker-core==1.0.52
schema==0.7.7
shellingham==1.5.4
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
starlette==0.47.2
tabulate==0.9.0
tblib==3.1.0
together==1.5.25
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zipp==3.23.0
